Renamed InspectorView.vue to InspectorViews.vue
