+++ src/ui/inspector/InspectorViews.vue
